+++ torch/_dynamo/polyfills/loader.py
@@ -1,0 +1,3 @@
+# Used to load and initialize polyfill handlers when importing torch._dynamo
+
+from . import builtins, functools, itertools, os  # noqa: F401
